add hosts appending with ENV variable
Rails 6 introduces some DNS rebinding protection, which allows you to
allowlist the domains that your app can be reached on. Other domains
that route to your app would be blocked when the request comes in with
the wrong host header.

Now, hosts other than `localhost` are not allowed and must be added to
the rails configuration via config.hosts <<

https://prathamesh.tech/2019/09/02/dns-rebinding-attacks-protection-in-rails-6/